// src/modules/wms/backend/types/wms-movement-order-detail.types.ts
// Auto-generated by Codegen Engine
export interface WmsMovementOrderDetailDO {
  id: string
  movement_order_id: string
  item_id: string
  qty: number
  tenant_id: string
  create_time: string
}

export interface WmsMovementOrderDetailVO {
  id: string
  movement_order_id: string
  item_id: string
  qty: number
  tenant_id: string
  create_time: string
}

export interface WmsMovementOrderDetailCreateDTO {
  movement_order_id: string
  item_id: string
  qty: number
}

export interface WmsMovementOrderDetailUpdateDTO extends Partial<WmsMovementOrderDetailCreateDTO> {
  id: string
}

export interface WmsMovementOrderDetailPageQuery {
  page?: number
  pageSize?: number
  movement_order_id?: string
  item_id?: string
}